simplifying Segment hierarchy
